原订单消息查询接口作废

+++ AsNum.Aliexpress.API/Entity/Message.cs
@@ -9,7 +9,7 @@ namespace AsNum.Xmj.API.Entity {
     public class Message {
 
         [JsonProperty("id")]
-        public int ID {
+        public long ID {
             get;
             set;
         }
